[Feat] 기록이 없을시 보여지는 HomeGuideVC x버튼으로로 dismiss 가능하게 구현 #131
